Add unit tests for Keras and PyTorch training functions; include mock data generators and validation checks
